+++ src/main/java/com/bbangle/bbangle/admin/admin/service/AdminAuthService.java
@@ -1,15 +1,15 @@
 package com.bbangle.bbangle.admin.admin.service;
 
-import com.bbangle.bbangle.admin.domain.Admin;
-import com.bbangle.bbangle.admin.admin.dto.AdminRequest;
 import com.bbangle.bbangle.admin.admin.dto.AdminLoginResponse;
+import com.bbangle.bbangle.admin.admin.dto.AdminRequest;
+import com.bbangle.bbangle.admin.domain.Admin;
 import com.bbangle.bbangle.admin.repository.AdminRepository;
 import com.bbangle.bbangle.common.redis.repository.RefreshTokenRepository;
 import com.bbangle.bbangle.common.role.Role;
+import com.bbangle.bbangle.config.security.jwt.TokenProvider;
 import com.bbangle.bbangle.exception.BbangleErrorCode;
 import com.bbangle.bbangle.exception.BbangleException;
 import com.bbangle.bbangle.token.domain.RefreshToken;
-import com.bbangle.bbangle.token.jwt.TokenProvider;
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -33,22 +33,22 @@ public class AdminAuthService {
     @Transactional
     public AdminLoginResponse login(AdminRequest.AdminLoginRequest request) {
         Admin admin = adminRepository.findByAccountId(request.accountId())
-            .orElseThrow(() -> new BbangleException(BbangleErrorCode.ADMIN_NOT_FOUND));
+                .orElseThrow(() -> new BbangleException(BbangleErrorCode.ADMIN_NOT_FOUND));
 
-        if(!passwordEncoder.matches(request.password(), admin.getPassword())) {
+        if (!passwordEncoder.matches(request.password(), admin.getPassword())) {
             throw new BbangleException(BbangleErrorCode.ADMIN_INVALID_PASSWORD);
         }
 
         String accessToken = tokenProvider.generateToken(admin.getId(), Role.ROLE_ADMIN,
-            ACCESS_TOKEN_DURATION);
+                ACCESS_TOKEN_DURATION);
         String refreshTokenVal = tokenProvider.generateToken(admin.getId(), Role.ROLE_ADMIN,
-            REFRESH_TOKEN_DURATION);
+                REFRESH_TOKEN_DURATION);
 
         RefreshToken refreshToken = refreshTokenRepository.findByAdminId(admin.getId())
-            .orElse(RefreshToken.builder()
-                .adminId(admin.getId())
-                .refreshToken(refreshTokenVal)
-                .build());
+                .orElse(RefreshToken.builder()
+                        .adminId(admin.getId())
+                        .refreshToken(refreshTokenVal)
+                        .build());
 
         // upsert 방식
         // 매 로그인마다 refresh token 갱신
@@ -56,9 +56,9 @@ public class AdminAuthService {
         refreshTokenRepository.save(refreshToken);
 
         return AdminLoginResponse.builder()
-            .accessToken(accessToken)
-            .refreshToken(refreshTokenVal)
-            .build();
+                .accessToken(accessToken)
+                .refreshToken(refreshTokenVal)
+                .build();
     }
 
     @Transactional
