+++ src/main/java/com/bbangle/bbangle/token/service/TokenService.java
@@ -1,8 +1,8 @@
 package com.bbangle.bbangle.token.service;
 
+import com.bbangle.bbangle.config.security.jwt.TokenProvider;
 import com.bbangle.bbangle.member.customer.service.MemberService;
 import com.bbangle.bbangle.member.domain.Member;
-import com.bbangle.bbangle.token.jwt.TokenProvider;
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,7 +17,7 @@ public class TokenService {
 
     public String createNewAccessToken(String refreshToken) {
         Long memberId = refreshTokenService.findByRefreshToken(refreshToken)
-            .getMemberId();
+                .getMemberId();
         Member member = memberService.findById(memberId);
         return tokenProvider.generateToken(member.getId(), member.getRole(), Duration.ofHours(2));
     }
