+++ src/test/java/com/bbangle/bbangle/notification/customer/controller/NotificationControllerSliceTest.java
@@ -19,14 +19,14 @@ import com.bbangle.bbangle.common.adaptor.slack.TestSlackAdaptorConfig;
 import com.bbangle.bbangle.common.page.NotificationCustomPage;
 import com.bbangle.bbangle.common.service.ResponseService;
 import com.bbangle.bbangle.config.JsonDataEncoder;
+import com.bbangle.bbangle.config.security.jwt.TestJwtPropertiesConfig;
+import com.bbangle.bbangle.config.security.jwt.TokenProvider;
 import com.bbangle.bbangle.exception.BbangleException;
 import com.bbangle.bbangle.exception.GlobalControllerAdvice;
 import com.bbangle.bbangle.notification.customer.dto.NotificationDetailResponseDto;
 import com.bbangle.bbangle.notification.customer.dto.NotificationResponse;
 import com.bbangle.bbangle.notification.customer.dto.NotificationUploadRequest;
 import com.bbangle.bbangle.notification.customer.service.NotificationService;
-import com.bbangle.bbangle.token.jwt.TestJwtPropertiesConfig;
-import com.bbangle.bbangle.token.jwt.TokenProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -41,11 +41,11 @@ import org.springframework.test.web.servlet.MockMvc;
 
 @DisplayName("[컨트롤러] NotificationController")
 @Import({
-    TestSlackAdaptorConfig.class,
-    JsonDataEncoder.class,
-    TokenProvider.class,
-    TestJwtPropertiesConfig.class,
-    ResponseService.class
+        TestSlackAdaptorConfig.class,
+        JsonDataEncoder.class,
+        TokenProvider.class,
+        TestJwtPropertiesConfig.class,
+        ResponseService.class
 })
 @WebMvcTest(controllers = NotificationController.class)
 class NotificationControllerSliceTest {
@@ -69,25 +69,25 @@ class NotificationControllerSliceTest {
         // Given
         Long cursorId = 1L;
         NotificationResponse response = new NotificationResponse(1L, "title1", "content1",
-            "2023-10-01 12:00");
+                "2023-10-01 12:00");
         NotificationCustomPage<List<NotificationResponse>> page =
-            NotificationCustomPage.from(List.of(response), 2L);
+                NotificationCustomPage.from(List.of(response), 2L);
         given(notificationService.getList(cursorId)).willReturn(page);
 
         // When & Then
         mvc.perform(get("/api/v1/notification")
-                .param("cursorId", cursorId.toString())
-                .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.success").value(true))
-            .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
-            .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()))
-            .andExpect(jsonPath("$.result.content[0].id").value(1))
-            .andExpect(jsonPath("$.result.content[0].title").value(response.title()))
-            .andExpect(jsonPath("$.result.content[0].content").value(response.content()))
-            .andExpect(jsonPath("$.result.content[0].createdAt").value(response.createdAt()))
-            .andExpect(jsonPath("$.result.nextCursor").value(page.getNextCursor()))
-            .andExpect(jsonPath("$.result.hasNext").value(page.getHasNext()));
+                        .param("cursorId", cursorId.toString())
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.success").value(true))
+                .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
+                .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()))
+                .andExpect(jsonPath("$.result.content[0].id").value(1))
+                .andExpect(jsonPath("$.result.content[0].title").value(response.title()))
+                .andExpect(jsonPath("$.result.content[0].content").value(response.content()))
+                .andExpect(jsonPath("$.result.content[0].createdAt").value(response.createdAt()))
+                .andExpect(jsonPath("$.result.nextCursor").value(page.getNextCursor()))
+                .andExpect(jsonPath("$.result.hasNext").value(page.getHasNext()));
         then(notificationService).should(times(1)).getList(cursorId);
         then(responseService).should(times(1)).getSingleResult(page);
     }
@@ -98,20 +98,20 @@ class NotificationControllerSliceTest {
         // Given
         Long id = 1L;
         NotificationDetailResponseDto detailResponse = new NotificationDetailResponseDto(
-            id, "title1", "content1", "2023-10-01 12:00");
+                id, "title1", "content1", "2023-10-01 12:00");
         given(notificationService.getNoticeDetail(id)).willReturn(detailResponse);
 
         // When & Then
         mvc.perform(get("/api/v1/notification/{id}", id)
-                .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.success").value(true))
-            .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
-            .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()))
-            .andExpect(jsonPath("$.result.id").value(id))
-            .andExpect(jsonPath("$.result.title").value(detailResponse.title()))
-            .andExpect(jsonPath("$.result.content").value(detailResponse.content()))
-            .andExpect(jsonPath("$.result.createdAt").value(detailResponse.createdAt()));
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.success").value(true))
+                .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
+                .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()))
+                .andExpect(jsonPath("$.result.id").value(id))
+                .andExpect(jsonPath("$.result.title").value(detailResponse.title()))
+                .andExpect(jsonPath("$.result.content").value(detailResponse.content()))
+                .andExpect(jsonPath("$.result.createdAt").value(detailResponse.createdAt()));
         then(notificationService).should(times(1)).getNoticeDetail(id);
         then(responseService).should(times(1)).getSingleResult(detailResponse);
     }
@@ -119,23 +119,23 @@ class NotificationControllerSliceTest {
     @DisplayName("Notification 상세 조회 API - 실패(존재하지 않는 ID)")
     @Test
     void givenNonExistingNotificationId_whenGetNoticeDetail_thenReturns4xxClientError()
-        throws Exception {
+            throws Exception {
         // Given
         Long nonExistingId = -1L;
         given(notificationService.getNoticeDetail(nonExistingId)).willThrow(
-            new BbangleException(NOTIFICATION_NOT_FOUND));
+                new BbangleException(NOTIFICATION_NOT_FOUND));
 
         // When & Then
         mvc.perform(get("/api/v1/notification/{id}", nonExistingId)
-                .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().is4xxClientError())
-            .andExpect(jsonPath("$.success").value(false))
-            .andExpect(jsonPath("$.code").value(NOTIFICATION_NOT_FOUND.getCode()))
-            .andExpect(jsonPath("$.message").value(NOTIFICATION_NOT_FOUND.getMessage()));
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().is4xxClientError())
+                .andExpect(jsonPath("$.success").value(false))
+                .andExpect(jsonPath("$.code").value(NOTIFICATION_NOT_FOUND.getCode()))
+                .andExpect(jsonPath("$.message").value(NOTIFICATION_NOT_FOUND.getMessage()));
 
         then(notificationService).should(times(1)).getNoticeDetail(nonExistingId);
         then(globalControllerAdvice).should(times(1))
-            .handleBbangleException(any(HttpServletRequest.class), any(BbangleException.class));
+                .handleBbangleException(any(HttpServletRequest.class), any(BbangleException.class));
         then(responseService).should(times(1)).getFailResult(anyString(), anyInt());
     }
 
@@ -144,18 +144,18 @@ class NotificationControllerSliceTest {
     void givenNotificationUploadRequest_whenUpload_thenReturnsSuccess() throws Exception {
         // Given
         NotificationUploadRequest uploadRequest = new NotificationUploadRequest("title1",
-            "content1");
+                "content1");
         willDoNothing().given(notificationService).upload(any(NotificationUploadRequest.class));
 
         // When & Then
         mvc.perform(post("/api/v1/notification")
-                .with(user("user").roles("USER"))
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(jsonDataEncoder.encode(uploadRequest)))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.success").value(true))
-            .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
-            .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()));
+                        .with(user("user").roles("USER"))
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(jsonDataEncoder.encode(uploadRequest)))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.success").value(true))
+                .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
+                .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()));
         then(notificationService).should(times(1)).upload(any(NotificationUploadRequest.class));
         then(responseService).should(times(1)).getSuccessResult();
     }
@@ -165,13 +165,13 @@ class NotificationControllerSliceTest {
     void givenNoAuthority_whenUploadNotification_thenReturns4xxClientError() throws Exception {
         // Given
         NotificationUploadRequest uploadRequest = new NotificationUploadRequest("title1",
-            "content1");
+                "content1");
 
         // When & Then
         mvc.perform(post("/api/v1/notification")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(jsonDataEncoder.encode(uploadRequest)))
-            .andExpect(status().is4xxClientError());
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(jsonDataEncoder.encode(uploadRequest)))
+                .andExpect(status().is4xxClientError());
     }
 
 }
